scripts/requirements.txt
pikepdf==10.5.1
Pillow>=10.4,<12
